fix: изменение наименований переменных под тесты

+++ src/main/java/ru/yandex/practicum/filmorate/storage/feed/FeedStorage.java
@@ -4,12 +4,11 @@ import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.enums.EventType;
 import ru.yandex.practicum.filmorate.model.enums.Operation;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 public interface FeedStorage {
 
     Collection<Feed> getFeedByUserId(Integer id);
 
-    void addFeed(Integer entityId, Integer userId, Integer timeStamp, EventType eventType, Operation operation);
+    void addFeed(Integer entityId, Integer userId, long timeStamp, EventType eventType, Operation operation);
 }
